fix: upgrade Astro to v4.12.3 to resolve props issue in dynamic routes
Upgraded Astro from v4.11.6 to v4.12.3 due to a problem with using props in dynamic routes as described in the "Astro Tutorial: Use props in dynamic routes".

The issue was related to a warning in the Terminal Console I overlooked about `export const prerender = true;`, which caused `Astro.props` to not work correctly.

commit-id:4e538fd6